LayoutTests/http/tests/plugins/cross-frame-object-access-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/plugins/cross-frame-object-access.html from frame with URL http://localhost:8000/plugins/resources/cross-frame-object-access.html. Domains, protocols and ports must match.



--------
Frame: 'childFrame'
--------

This tests that plug-ins can access objects in other frames as allowed by the security model enforced in WebCore.
SUCCESS
